move pkg entity, repo into dao
