+++ module11/src/main/java/com/example/module11/dao/repo/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.module11.repo;
+package com.example.module11.dao.repo;
 
-import com.example.module11.entity.UserEntity;
+import com.example.module11.dao.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
